refactor: move config file to toml, using toml_edit to preserve user changes to config moving foreward.

+++ crates/config-lib/src/migrate/m2t3/mod.rs
@@ -1,5 +1,5 @@
-pub const PREV_CONFIG_VERSION: u16 = 2;
-pub const NEXT_CONFIG_VERSION: u16 = 3;
+pub const PREV_CONFIG_VERSION: u64 = 2;
+pub const NEXT_CONFIG_VERSION: u64 = 3;
 mod convert;
 mod old_structs;
